+++ packages/scaffold/src/modal/w3m-onramp-widget/index.ts
@@ -78,8 +78,8 @@ export class W3mOnrampWidget extends LitElement {
             ${BUY_PRESET_AMOUNTS.map(
               amount =>
                 html`<wui-button
-                  variant=${this.paymentAmount === amount ? 'accentBg' : 'shade'}
-                  size="xs"
+                  variant=${this.paymentAmount === amount ? 'accent' : 'neutral'}
+                  size="md"
                   textVariant="paragraph-600"
                   fullWidth
                   @click=${() => this.selectPresetAmount(amount)}
@@ -99,7 +99,7 @@ export class W3mOnrampWidget extends LitElement {
     return this.connected
       ? html`<wui-button
           @click=${this.getQuotes.bind(this)}
-          variant="fill"
+          variant="main"
           fullWidth
           size="lg"
           borderRadius="xs"
@@ -108,7 +108,7 @@ export class W3mOnrampWidget extends LitElement {
         </wui-button>`
       : html`<wui-button
           @click=${this.openModal.bind(this)}
-          variant="accentBg"
+          variant="accent"
           fullWidth
           size="lg"
           borderRadius="xs"