+++ packages/ui/src/composites/wui-list-social/styles.ts
@@ -6,7 +6,7 @@ export default css`
     padding: 7px var(--wui-spacing-l) 7px var(--wui-spacing-xs);
     width: 100%;
     justify-content: flex-start;
-    background-color: var(--wui-gray-glass-002);
+    background-color: var(--wui-color-gray-glass-002);
     border-radius: var(--wui-border-radius-xs);
     color: var(--wui-color-fg-100);
   }
@@ -32,7 +32,7 @@ export default css`
   }
 
   button:disabled {
-    background-color: var(--wui-gray-glass-015);
-    color: var(--wui-gray-glass-015);
+    background-color: var(--wui-color-gray-glass-015);
+    color: var(--wui-color-gray-glass-015);
   }
 `