+++ packages/ui/src/composites/wui-profile-button/index.ts
@@ -21,7 +21,7 @@ export class WuiProfileButton extends LitElement {
 
   @property() public avatarSrc?: string = undefined
 
-  @property({ type: Boolean }) public isProfileName = false
+  @property() public profileName?: string = ''
 
   @property() public address = ''
 
@@ -40,10 +40,10 @@ export class WuiProfileButton extends LitElement {
         <wui-flex gap="xs" alignItems="center">
           <wui-text variant="large-600" color="fg-100">
             ${UiHelperUtil.getTruncateString({
-              string: this.address,
-              charsStart: this.isProfileName ? 18 : 4,
-              charsEnd: this.isProfileName ? 0 : 4,
-              truncate: this.isProfileName ? 'end' : 'middle'
+              string: this.profileName || this.address,
+              charsStart: this.profileName ? 18 : 4,
+              charsEnd: this.profileName ? 0 : 4,
+              truncate: this.profileName ? 'end' : 'middle'
             })}
           </wui-text>
           <wui-icon size="sm" color="fg-200" name=${this.icon}></wui-icon>