+++ packages/scaffold/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @web3modal/scaffold
 
+## 4.2.1
+
+### Patch Changes
+
+- Hotfix to support injected and announced wallets in in app browsers
+
+- Updated dependencies []:
+  - @web3modal/common@4.2.1
+  - @web3modal/core@4.2.1
+  - @web3modal/siwe@4.2.1
+  - @web3modal/ui@4.2.1
+  - @web3modal/wallet@4.2.1
+
 ## 4.2.0
 
 ### Patch Changes